+++ bin/imdb_parser.rb
@@ -6,6 +6,6 @@ $LOAD_PATH.unshift(File.expand_path('lib'))
 require 'movies'
 
 collection = Movies::MovieCollection.new
-parser = Movies::ImdbParser.new
+parser = Movies::Utils::ImdbParser.new
 parser.run(collection)
 parser.save
