+++ lib/movies.rb
@@ -5,20 +5,13 @@ require 'money'
 require 'virtus'
 require 'movies/movie'
 require 'movies/movie_collection'
-require 'movies/country_selection'
-require 'movies/genre_selection'
 require 'movies/ancient_movie'
 require 'movies/classic_movie'
 require 'movies/modern_movie'
 require 'movies/new_movie'
-require 'movies/cashbox'
-require 'movies/hall'
-require 'movies/period'
 require 'movies/cinema'
-require 'movies/netflix'
-require 'movies/theatre'
-require 'movies/haml_presenter'
-require 'movies/imdb_parser'
-require 'movies/tmdb_parser'
+require 'movies/utils/haml_presenter'
+require 'movies/utils/imdb_parser'
+require 'movies/utils/tmdb_parser'
 
 I18n.enforce_available_locales = false
