+++ lib/movies/movie_collection.rb
@@ -5,10 +5,11 @@ require 'csv'
 require 'yaml'
 
 module Movies
-  class MovieCollection
-    BUDGETS_FILE = 'budgets.yml'
-    DEFAULT_PATH = 'movies.txt'
-    INFO_FILE = 'info.yml'
+  class MovieCollection    
+    DATA_FOLDER = 'data'
+    BUDGETS_FILE = "#{DATA_FOLDER}/budgets.yml"
+    DEFAULT_PATH = "#{DATA_FOLDER}/movies.txt"
+    INFO_FILE = "#{DATA_FOLDER}/info.yml"
 
     include Enumerable
