+++ spec/utils/imdb_parser_spec.rb
@@ -1,26 +1,26 @@
 # coding: utf-8
 # frozen_string_literal: true
 
-describe Movies::TmdbParser do
+describe Movies::Utils::ImdbParser do
   let(:collection) { Movies::MovieCollection.new.first(1) }
   let(:parser) { described_class.new }
-  let(:info) { {'tt0111161' => {title:'Побег из Шоушенка', poster_path: '/sRBNv6399ZpCE4RrM8tRsDLSsaG.jpg'}} }
+  let(:info) { {'tt0111161' => '$25,000,000'} }
 
   describe '#run' do
     subject { parser.run(collection) }
 
     it 'get movie attributes as result' do
-      VCR.use_cassette('movies') do
-        expect(subject).to eq(info)
+      VCR.use_cassette('page') do
+        expect { subject }.to change(parser, :budgets).from({}).to(info)
       end
     end
   end
 
   describe '#save' do
-    subject { parser.save(info) }
+    subject { parser.save }
 
     it 'saves result to local file' do
-      expect(File).to receive(:write).with('info.yml', an_instance_of(String))
+      expect(File).to receive(:write).with('data/budgets.yml', an_instance_of(String))
       subject
     end
   end
